+++ src/main/java/com/robertx22/age_of_exile/capability/bases/CapSyncUtil.java
@@ -10,7 +10,6 @@ public class CapSyncUtil {
     public static void syncAll(PlayerEntity player) {
         syncEntityCap(player);
         syncSpells(player);
-        syncStatPoints(player);
         syncFavor(player);
     }
 
@@ -22,11 +21,6 @@ public class CapSyncUtil {
     public static void syncEntityCap(PlayerEntity player) {
         EntityCap.UnitData data = Load.Unit(player);
         data.syncToClient(player);
-    }
-
-    public static void syncStatPoints(PlayerEntity player) {
-        Load.statPoints(player)
-            .syncToClient(player);
     }
 
     public static void syncFavor(PlayerEntity player) {