+++ src/main/java/com/robertx22/age_of_exile/database/data/stats/types/core_stats/Dexterity.java
@@ -4,7 +4,6 @@ import com.robertx22.age_of_exile.database.OptScaleExactStat;
 import com.robertx22.age_of_exile.database.data.stats.types.core_stats.base.BaseCoreStat;
 import com.robertx22.age_of_exile.database.data.stats.types.defense.DodgeRating;
 import com.robertx22.age_of_exile.database.data.stats.types.offense.Accuracy;
-import com.robertx22.age_of_exile.database.data.stats.types.offense.AttackStyleDamage;
 import com.robertx22.age_of_exile.uncommon.enumclasses.ModType;
 
 import java.util.Arrays;
@@ -18,8 +17,7 @@ public class Dexterity extends BaseCoreStat {
     private Dexterity() {
         super(Arrays.asList(
             new OptScaleExactStat(20, 20, Accuracy.getInstance(), ModType.FLAT),
-            new OptScaleExactStat(0.25F, 0.25F, AttackStyleDamage.RANGED, ModType.FLAT),
-            new OptScaleExactStat(1, 1, DodgeRating.getInstance(), ModType.LOCAL_INCREASE)
+            new OptScaleExactStat(2, 2, DodgeRating.getInstance(), ModType.LOCAL_INCREASE)
         ));
     }