+++ src/main/java/com/robertx22/age_of_exile/database/data/stats/types/core_stats/Strength.java
@@ -3,8 +3,7 @@ package com.robertx22.age_of_exile.database.data.stats.types.core_stats;
 import com.robertx22.age_of_exile.database.OptScaleExactStat;
 import com.robertx22.age_of_exile.database.data.stats.types.core_stats.base.BaseCoreStat;
 import com.robertx22.age_of_exile.database.data.stats.types.defense.Armor;
-import com.robertx22.age_of_exile.database.data.stats.types.offense.Accuracy;
-import com.robertx22.age_of_exile.database.data.stats.types.offense.AttackStyleDamage;
+import com.robertx22.age_of_exile.database.data.stats.types.resources.health.Health;
 import com.robertx22.age_of_exile.uncommon.enumclasses.ModType;
 
 import java.util.Arrays;
@@ -16,9 +15,8 @@ public class Strength extends BaseCoreStat {
 
     private Strength() {
         super(Arrays.asList(
-            new OptScaleExactStat(10, 10, Accuracy.getInstance(), ModType.FLAT),
-            new OptScaleExactStat(0.5F, 0.5F, AttackStyleDamage.MELEE, ModType.FLAT),
-            new OptScaleExactStat(1, 1, Armor.getInstance(), ModType.LOCAL_INCREASE)
+            new OptScaleExactStat(5, 5, Health.getInstance(), ModType.FLAT),
+            new OptScaleExactStat(2, 2, Armor.getInstance(), ModType.LOCAL_INCREASE)
         ));
     }