+++ src/main/java/com/robertx22/age_of_exile/mmorpg/registers/common/C2SPacketRegister.java
@@ -1,7 +1,6 @@
 package com.robertx22.age_of_exile.mmorpg.registers.common;
 
 import com.robertx22.age_of_exile.vanilla_mc.packets.ModifyItemPacket;
-import com.robertx22.age_of_exile.vanilla_mc.packets.SpendStatPointsPacket;
 import com.robertx22.age_of_exile.vanilla_mc.packets.perks.PerkChangePacket;
 import com.robertx22.age_of_exile.vanilla_mc.packets.registry.RequestRegistriesPacket;
 import com.robertx22.age_of_exile.vanilla_mc.packets.spells.HotbarSetupPacket;
@@ -15,7 +14,6 @@ public class C2SPacketRegister {
 
         Packets.registerClientToServerPacket(new ModifyItemPacket());
         Packets.registerClientToServerPacket(new RequestSyncCapToClient());
-        Packets.registerClientToServerPacket(new SpendStatPointsPacket());
         Packets.registerClientToServerPacket(new HotbarSetupPacket());
         Packets.registerClientToServerPacket(new TellServerToCastSpellPacket());
         Packets.registerClientToServerPacket(new PerkChangePacket());