+++ src/main/java/com/robertx22/age_of_exile/mmorpg/registers/common/ComponentRegisters.java
@@ -2,7 +2,10 @@ package com.robertx22.age_of_exile.mmorpg.registers.common;
 
 import com.robertx22.age_of_exile.capability.entity.EntityCap;
 import com.robertx22.age_of_exile.capability.entity.EntityPerks;
-import com.robertx22.age_of_exile.capability.player.*;
+import com.robertx22.age_of_exile.capability.player.PlayerDeathData;
+import com.robertx22.age_of_exile.capability.player.PlayerFavor;
+import com.robertx22.age_of_exile.capability.player.PlayerSkills;
+import com.robertx22.age_of_exile.capability.player.PlayerSpellCap;
 import com.robertx22.age_of_exile.capability.world.WorldAreas;
 import com.robertx22.age_of_exile.mmorpg.Ref;
 import nerdhub.cardinal.components.api.ComponentRegistry;
@@ -28,12 +31,6 @@ public class ComponentRegisters {
             new Identifier(Ref.MODID, "spells"),
             PlayerSpellCap.ISpellsCap.class)
             .attach(EntityComponentCallback.event(PlayerEntity.class), x -> new PlayerSpellCap.DefaultImpl());
-
-    public ComponentType<PlayerStatsCap.IPlayerStatPointsData> PLAYER_STAT_POINTS =
-        ComponentRegistry.INSTANCE.registerIfAbsent(
-            new Identifier(Ref.MODID, "stat_points"),
-            PlayerStatsCap.IPlayerStatPointsData.class)
-            .attach(EntityComponentCallback.event(PlayerEntity.class), x -> new PlayerStatsCap.DefaultImpl());
 
     public ComponentType<PlayerFavor> PLAYER_FAVOR =
         ComponentRegistry.INSTANCE.registerIfAbsent(
@@ -67,7 +64,6 @@ public class ComponentRegisters {
 
         EntityComponents.setRespawnCopyStrategy(UNIT_DATA, RespawnCopyStrategy.ALWAYS_COPY);
         EntityComponents.setRespawnCopyStrategy(PLAYER_SPELLS, RespawnCopyStrategy.ALWAYS_COPY);
-        EntityComponents.setRespawnCopyStrategy(PLAYER_STAT_POINTS, RespawnCopyStrategy.ALWAYS_COPY);
         EntityComponents.setRespawnCopyStrategy(PERKS, RespawnCopyStrategy.ALWAYS_COPY);
         EntityComponents.setRespawnCopyStrategy(PLAYER_FAVOR, RespawnCopyStrategy.ALWAYS_COPY);
         EntityComponents.setRespawnCopyStrategy(PLAYER_SKILLS, RespawnCopyStrategy.ALWAYS_COPY);