+++ src/main/java/com/robertx22/age_of_exile/mmorpg/registers/common/items/MiscItemsRegistrator.java
@@ -64,7 +64,6 @@ public class MiscItemsRegistrator extends BaseItemRegistrator {
     public Item GOLDEN_ORB = item(new SimpleMatItem(), "mat/golden_orb");
     public Item MYTHIC_ESSENCE = item(new SimpleMatItem(), "mat/mythic_essence");
 
-    public ResetStatPointsItem RESET_STATS_POTION = item(new ResetStatPointsItem());
     public ResetAllPerksItem RESET_ALL_PERKS = item(new ResetAllPerksItem());
     public ResetAllSpellsItem RESET_ALL_SPELLS = item(new ResetAllSpellsItem());
     public AddResetPerkPointsItem ADD_RESET_PERK_POINTS = item(new AddResetPerkPointsItem());